Fix DMA and remove buffer size configuration
- Whilst working on the python code and trying to maintain a level
of compatibility with the PixelBlaze output expander, I realised
that we don't really need the user to configure the size of the buffer
ahead of time. They can just send us pixel data and as long as it fits
inside the fixed-sized buffer, we can send it out. So I removed the
storing and configuring of the channel output size.

- There was an issue with the DMA because the DMA was being pointed
at the USB recieve buffer, even after we'd carefully memcpy'd it to a
per-channel buffer. So the data in the buffer was being trampled by the
next USB transfer whilst the DMA was happening. Also fixed it to use
dma_encode_transfer_count as per the documentation.